03.07.2019 -0-

+++ src/lesson35/controller/HotelController.java
@@ -1,7 +1,10 @@
 package lesson35.controller;
 
 import lesson35.model.Hotel;
+import lesson35.repository.UserDAO;
 import lesson35.service.HotelService;
+
+import java.nio.file.AccessDeniedException;
 
 public class HotelController {
     private HotelService hotelService = new HotelService();
@@ -10,6 +13,8 @@ public class HotelController {
     only for administrators
      */
     public void addHotel(Hotel hotel) throws Exception {
+        if (!UserDAO.isAdmin())
+            throw new AccessDeniedException("User do not have permission");
         hotelService.addHotel(hotel);
     }
 
@@ -17,6 +22,8 @@ public class HotelController {
     only for administrators
      */
     public void deleteHotel(long hotelId) throws Exception {
+        if (!UserDAO.isAdmin())
+            throw new AccessDeniedException("User do not have permission");
         hotelService.deleteHotel(hotelId);
     }
