+++ src/lesson35/controller/RoomController.java
@@ -2,8 +2,10 @@ package lesson35.controller;
 
 import lesson35.model.Filter;
 import lesson35.model.Room;
+import lesson35.repository.UserDAO;
 import lesson35.service.RoomService;
 
+import java.nio.file.AccessDeniedException;
 import java.util.ArrayList;
 
 public class RoomController {
@@ -13,6 +15,8 @@ public class RoomController {
     only for administrators
      */
     public void addRoom(Room room) throws Exception {
+        if (!UserDAO.isAdmin())
+            throw new AccessDeniedException("User do not have permission");
         roomService.addRoom(room);
     }
 
@@ -20,9 +24,14 @@ public class RoomController {
     only for administrators
      */
     public void deleteRoom(long roomId) throws Exception {
+        if (!UserDAO.isAdmin())
+            throw new AccessDeniedException("User do not have permission");
         roomService.deleteRoom(roomId);
     }
 
+    /*
+    for users
+    */
     public ArrayList<Room> findRooms(Filter filter) {
         if (filter != null)
             return roomService.findRooms(filter);
